Add photo gallery to reviews (#257)
* drag-and-drop

* isSubmitting

* submit photo

* submit review last

* rename

* move Segment out

* remove user prop

* get user

* remove unused

* show gallery
