+++ app/src/components/UI/Organisms/ReviewForm/ReviewForm.tsx
@@ -60,7 +60,7 @@ export const ReviewForm = (props: Props) => {
           value={textAreaInput}
           onChange={handleChange}
         />
-        <p>Attach pictures</p>
+        <p>Attach pictures (Must be .jpg)</p>
         <Dropzone
           accept="image/jpeg"
           onDropAccepted={onDrop}
