+++ app/src/components/UI/Pages/Vendor.tsx
@@ -11,7 +11,7 @@ import {
   Photo,
   useS3CredentialsMutation,
 } from "../../../api";
-import { Container, Divider, Grid, Header } from "semantic-ui-react";
+import { Container, Divider, Grid, Header, Segment } from "semantic-ui-react";
 import VendorDetailCards from "../Atoms/VendorDetailCards/VendorDetailCards";
 import { Review } from "../Organisms/Review/Review";
 import { ReviewForm } from "../Organisms/ReviewForm/ReviewForm";
@@ -19,7 +19,7 @@ import { v4 as uuid } from "uuid";
 import { useAppSelector } from "../../../store";
 import { DateTime } from "luxon";
 import Buttons from "../Atoms/Button/Buttons";
-import Gallery from "../Organisms/VendorGallery/VendorGallery";
+import Gallery from "../Organisms/VendorGallery/Gallery";
 import styles from "./vendor.module.css";
 import { uploadToS3 } from "../../../aws";
 
@@ -33,8 +33,7 @@ export function Vendor(): React.ReactElement {
   const reviews = reviewsQuery.data;
   const [submitReview] = useCreateReviewMutation();
   const token = useAppSelector((state) => state.token.token);
-  const { data: photos, isSuccess: photosIsSuccess } =
-    usePhotosByLinkIDQuery(vendorID);
+  const { data: photos } = usePhotosByLinkIDQuery(vendorID);
   const navigate = useNavigate();
   const [isSubmitting, setIsSubmitting] = useState(false);
   const [createPhoto] = useCreatePhotoMutation();
@@ -96,7 +95,11 @@ export function Vendor(): React.ReactElement {
             <h1 className={styles.name}>{vendor?.Name}</h1>
           </Grid.Row>
           <Grid.Row>
-            {photosIsSuccess ? <Gallery photos={photos!} /> : null}
+            {photos ? (
+              <Segment style={{ width: "100%" }}>
+                <Gallery photos={photos} photoHeight={150} />
+              </Segment>
+            ) : null}
           </Grid.Row>
           <Grid.Row>
             <Grid.Column width={6}>
